Todo_AI_Chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0